engine: shorten Bullet's include directive
The top-level name of the dependency is specific to our project setup,
and we'd better not hardcode it (especially since by default when
cloning Bullet's github repo it gets stored in a directory named
"bullet3", and not "Bullet").

The Makefile_Linux_Editor already includes "External/Bullet", so the
prefix is not even needed.